Isolate legacy hybrid pattern allocation
